adding average and 90th percentile scores
This update includes the numpy library for calculating the mean and percentile. The "SDG_AVG" column will contain the average score of all SDG predictions for each row, and the "SDG_90th_percentile" column will contain the 90th percentile score. These additional metrics provide a broader understanding of the distribution and intensity of the SDG predictions for each text entry.